pyproject.toml
[project]
name = "lighthouse-exporter"
version = "0.1.0"
description = "CLI tool to export Google Lighthouse reports as Markdown via PageSpeed Insights API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
